pyproject.toml
[project]
name = "liberliber-rss-feed"
version = "0.1.0"
description = "LiberLiber rss feed"
readme = "README.md"
license = { text = "AGPL-3.0-or-later" }

authors = [
    { name = "Vito De Tullio", email = "vito.detullio@gmail.com" },
]
requires-python = "==3.12.*"
classifiers = [
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "defusedxml>=0.7.1",
    "mastodon-py>=1.8.1",
    "python-dotenv>=1.0.1",
]

[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
dev = [
    "ruff>=0.6.4",
    "mypy>=1.11.2",
    "pyproject-fmt>=2.2.3",
    "coverage>=7.6.1",
    "setuptools>=74.1.2",
    "types-defusedxml>=0.7.0.20240218",
]

[tool.pdm.scripts]
_.env = { PYTHONPATH = "src:tests" }
lint = { composite = [
    "pyproject-fmt pyproject.toml",
    "ruff format {args:.}",
    "ruff check --exit-zero {args:.}",
    "dmypy run {args:.}",
] }
start = 'src/__main__.py'
tests = { composite = [
    "coverage run -m unittest discover --verbose --locals --failfast --catch --start-directory {args:tests}",
    "coverage report",
    "coverage erase",
] }

[tool.ruff]
line-length = 80
src = [ "src", "tests" ]

fix = true
format.quote-style = "single"
format.skip-magic-trailing-comma = true
format.docstring-code-format = true
lint.select = [ "ALL" ]
lint.ignore = [
    "ANN1",
    "COM812", # conflicts with formatter
    "D1",     # disabled by me
    "D203",   # incompatibilities
    "D213",   # incompatibilities
    "ISC001", # incompatibilities
    "PT009",  # incompatibilities
    "PT027",  # disabled by me
    "Q001",   # incompatibilities
    "Q002",   # incompatibilities
]
lint.flake8-annotations.allow-star-arg-any = true
lint.flake8-quotes.docstring-quotes = "single"
lint.flake8-quotes.inline-quotes = "single"
lint.flake8-quotes.multiline-quotes = "single"
lint.flake8-type-checking.quote-annotations = true
lint.isort.force-single-line = true

[tool.pyproject-fmt]
column_width = 80
indent = 4

[tool.coverage.run]
branch = true
source = [ "src", "tests" ]
omit = [ "src/__main__.py" ]
[tool.coverage.report]
exclude_also = [
    "if TYPE_CHECKING:",
    "def main() -> None:",
]
show_missing = true
sort = "Cover"
skip_empty = true

[tool.mypy]
mypy_path = [ "src", "tests", "stubs" ]
strict = true
warn_unused_configs = true
